.travis.yml: Fix python dependencies for image signing
Previously, we only installed mbed-tools and python packages listed in
`mbed-os/tools/cmake/requirements.txt` which contains the bare minimum
for Mbed CLI 2. This is not sufficient, and some target-specific
dependencies are listed in the top-level `mbed-os/requirements.txt`,
so switch to install the latter.

Also take advantage of `mbed-tools deploy` which performs a shallow
clone of mbed-os.

Fixes #25